Add db:schema:load and db:seed to containers db setup
This fixes an error during docker-compose up where the short-lived
migration container would not be able to successfully run db:migrate.

Adding the rails commands to load the db schema and seed the db
ensures that when it runs db:migrate, it has the database setup as it
will need.