537: Bump lint parser from ecma 2017 to ecma 2018
This commit bumps up the linter's parser to ECMA 2018, which is
consistent with our current linting on ZAP.

The parser matters for linting because the linter has to know which
javascript rules we want it to use
(for example, will it recognize the
spread operator or not?).

Since the spread operator is no longer experimental in ecma 2018, I
also removed the node override rules that allowed the (then experimental)
rest operator. See here: https://eslint.org/docs/user-guide/migrating-to-5.0.0#experimental-object-rest-spread

Checked ember tests to see if this raised any lint errors - it didn't.