Remove ember-ast-hot-load
This addon is throwing a few very long warning messages in our test output. This is a convenience addon that “hot reloads” components during development. I’m not sure I want it in the source code right now unless anyone feels strongly about having it (it does not work that well with everything).