564: add test for user manual input of temporal mode split values
When a user is on the Manual tab and the Temporal Splits tab, they
can enter mode split values for each time of day (am, md, pm, saturday)
AND for each mode (e.g. auto, taxi, bus). The totals that are
calculated through the `total` computed property in modal-splits.js
produces the correct total value.