590 Remove Community Facility condition from detailed analysis trigger
Stakeholders clarified that the existence of a Community Facility
land use in a Project should not trigger a transportation
Detailed Analysis. This was previously the case, so this commit
removes that condition from the Detailed Analysis trigger.

This commit removes computed properties and UI features
related to the Community Facility trigger.

It also accordingly updates previous unit tests, and adds
one acceptance test.