484: Add 2019 public schools data package
This commit adds a 2019 public schools data package.

This uses the new public schools data package schema, and assumes
that the schema is updated for 2017 and 2018 as per issues #637 and #630

For the data package to take effect, someone must manually run
rails db:seed on the target database (local, staging, or production).