627 Add containerized script to load new acs table into ceqr_data
This commit introduces a command, `yarn run load-acs-data`, to load
a csv into the ceqr_data db's nyc_acs.2018 table. The csv
should be generated by the `yarn download-acs-data` command.

The command executes a Python script to facilitate connecting
to a DB and loading the table. Docker is used to run the script
in a controlled environment.

Though the two *-acs-data commands and scripts are written to
specifically download and load 2018 data, the code can be
tweaked for loading data for future years, assuming the
ACS REST API and CEQR nyc_acs schema stay unchanged.