659: Only load subdistricts when at Public Schools Analysis Existing Conditions step
This commit solves latency in opening and editing projects. It does so
by deferring loading of Public Schools Analysis' subdistricts_geojson
attribute. That attribute requires the backend to compute subdistrict
geometries by querying the Mappluto table, which turns out to be a slow
computation. Previously, the app loaded this attribute when a user
clicks to open a project. Now, this attribute only loads when the maps
load on the Existing Conditions or No Action steps of the analysis. It
introduces a loading spinner to the UI to indicate the subdistrict
map is loading.