Remove "% Funded" from sca schools under construction
Capital Planning confirmed that all the projects we have in this dataset
should be assumed funded enough to be included in ceqr public schools
analysis.

This commit removes the % Funded column from the schools under
construction table, and updates the language for the user in the
instructions and tooltip (removing "AND % funded is higher than 80%).
This language was confirmed by Capital Planning.