637: add migration that renames id hash key project_dsf to uid
Previously, project_dsf field was being used as a unique identifier to
check if certain projects had already been loaded. project_dsf was
removed as a field in the most recent public schools data update, and
the field uid was created to act as the unique identifier instead.
This commit renames the field `project_dsf` to `uid` on the
sca_projects column in public_schools_analyses table in the ceqr_rails
database for past projects that were created with the old schema.