Create 2019 public schools multipliers data
This commit adds the static JSON for the November 2019 SCA Public Schools multipliers.  It uses the same schema as november 2018.

We get “csd”, “ps”, “is”, “hs”, and “borocode” directly from SCA’s Capital Plan - Section 5:
http://www.nycsca.org/Community/Capital-Plan-Reports-Data#Local-Law-167-Reports-352

We generate the psisThreshold & hsThreshold by calculating:
studentThreshold / multiplier

For example, for csd 1, to calculate the number of housing units (the threshold for analysis), we take 50 (# of projected new students) and divide by (0.4+0.1), the sum of the ps and is student multipliers.  This would mean 1,000 housing units in this csd would project an additional 150 ps/is students, which triggers a CEQR public schools analysis.