Use new multipliers w/ new public schools data
This commit updates the logic for determining which multipliers version to use when the user is using the new november 2019 public schools data package.

On the backend, we serve the “november-2019.json” multpliers when the user is using the “november_2019” data package (or later in case we add a new data package in the future).

On the frontend, this commit configures the public schools analysis model to use the same logic for either “november-2018” or “november-2019” — since they have the same schema the frontend handling of it remains the same.