Use unrounded multipliers for preciseness
This commit updates the november 2019 multipliers to use the unrounded, raw multiplier values from SCA for primary and intermediate schools, and adjusts the frontend templates to display a rounded number.

Previous multiplier versions had calculated and hard-coded the housing thresholds from the unrounded student multipliers, but bring in a hard-coded rounded multiplier.

This means the analysis threshold calculation that happens to indicate whether or not a detailed analysis is needed (hits the threshold) for older multiplier verions uses the rounded multiplier value, but the hard-coded unit threshold in the table shows a precise calculation.

SCA’s data does not provide the HS multipliers unrounded.  Conversation with Capital Planning decided to use the unrounded multipliers for the calculations for PS and IS.

To maintain the UI display of a rounded multiplier, I used ember-math-helper to round the display in the templates.