Fix existing conditions table metadata
This commit fixes the source metadata for existing conditions.

The new computed wasn’t working because the dataPackage on public schools analysis is an async-loaded relationship — the fix was to wrap the invocation of the table in a conditional statement to ensure dataPackage is available.

Also updated the bluebook dates in the datapackage metadata.